+++ tests/test_signalfd.py
@@ -1,3 +1,4 @@
+import errno
 import os
 import signal
 
@@ -14,6 +15,25 @@ def test_block_and_read():
         assert si.ssi_signo == signal.SIGUSR1
     finally:
         os.close(fd)
+
+
+def test_read_fd_no_data():
+    fd = signalfd.signalfd(-1, [signal.SIGUSR1], signalfd.SFD_CLOEXEC | signalfd.SFD_NONBLOCK )
+    try:
+        err = pytest.raises(IOError, signalfd.read_siginfo, fd)
+        assert err.value.errno == errno.EAGAIN
+    finally:
+        os.close(fd)
+
+
+def test_read_fh_no_data():
+    fd = signalfd.signalfd(-1, [signal.SIGUSR1], signalfd.SFD_CLOEXEC | signalfd.SFD_NONBLOCK )
+    fh = os.fdopen(fd, 'rb')
+    try:
+        err = pytest.raises(IOError, signalfd.read_siginfo, fh)
+        assert err.value.errno == errno.EAGAIN
+    finally:
+        fh.close()
 
 
 def test_read_closed():